requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
lxml==5.1.0
multidict==6.0.4
PyExecJS==1.5.1
requests==2.31.0
six==1.16.0
urllib3==2.1.0
yarl==1.9.4
